feat: :sparkles: adding some sync functionality

+++ src/LeaderBoard/Infrastructure/Data/DataSeeder.cs
@@ -1,5 +1,6 @@
 using System.Globalization;
 using AutoBogus;
+using Humanizer;
 using LeaderBoard.Infrastructure.Data.EFContext;
 using LeaderBoard.Models;
 using LeaderBoard.SharedKernel.Data.Contracts;
@@ -29,7 +30,10 @@ public class DataSeeder : ISeeder
 
     public async Task SeedAsync()
     {
-        await DeleteAllKeys(_redisConnection, _redisDatabase);
+        if (_leaderBoardOptions.CleanupRedisOnStart)
+        {
+            await DeleteAllKeys(_redisConnection, _redisDatabase);
+        }
 
         var playerScores = new AutoFaker<PlayerScore>()
             .RuleFor(x => x.Country, f => f.Address.Country())
@@ -37,43 +41,52 @@ public class DataSeeder : ISeeder
             .RuleFor(x => x.LastName, f => f.Name.LastName())
             .RuleFor(x => x.PlayerId, f => f.Internet.UserName())
             .RuleFor(x => x.Score, f => f.Random.Number(1, 10000))
+            // we don't set rank here because for evaluating rank correctly we need all items present in database but here we have to add items one by one
             .RuleFor(x => x.Rank, f => null)
             .RuleFor(x => x.UpdatedAt, DateTime.Now)
             .RuleFor(x => x.CreatedAt, DateTime.Now)
+            .RuleFor(x => x.LeaderBoardName, Constants.GlobalLeaderBoard)
             .Generate(1000);
 
+        // for using cache-aside we need our database fill firstly
         if (!_leaderBoardOptions.UseReadCacheAside)
         {
-            foreach (var playerScore in playerScores)
+            if (_redisDatabase.SortedSetLength(Constants.GlobalLeaderBoard) != 0)
             {
-                // store the summary of our player-score in sortedset
-                await _redisDatabase.SortedSetAddAsync(
-                    Constants.GlobalLeaderBoard,
-                    playerScore.PlayerId,
-                    playerScore.Score
-                );
+                foreach (var playerScore in playerScores)
+                {
+                    var key = $"{nameof(PlayerScore).Underscore()}:{playerScore.PlayerId}";
 
-                // store detail of out score-player in hashset. it is related to its score information with their same unique identifier
-                await _redisDatabase.HashSetAsync(
-                    playerScore.PlayerId,
-                    new HashEntry[]
-                    {
-                        new(nameof(PlayerScore.Country).ToLower(), playerScore.Country),
-                        new(nameof(PlayerScore.FirstName).ToLower(), playerScore.FirstName),
-                        new(nameof(PlayerScore.LastName).ToLower(), playerScore.LastName),
-                        new(
-                            nameof(PlayerScore.UpdatedAt).ToLower(),
-                            playerScore.UpdatedAt.ToString(CultureInfo.InvariantCulture)
-                        ),
-                        new(
-                            nameof(PlayerScore.CreatedAt).ToLower(),
-                            playerScore.CreatedAt.ToString(CultureInfo.InvariantCulture)
-                        )
-                    }
-                );
+                    // store the summary of our player-score in sortedset
+                    await _redisDatabase.SortedSetAddAsync(
+                        Constants.GlobalLeaderBoard,
+                        key,
+                        playerScore.Score
+                    );
+
+                    // store detail of out score-player in hashset. it is related to its score information with their same unique identifier
+                    await _redisDatabase.HashSetAsync(
+                        key,
+                        new HashEntry[]
+                        {
+                            new(nameof(PlayerScore.Country).Underscore(), playerScore.Country),
+                            new(nameof(PlayerScore.FirstName).Underscore(), playerScore.FirstName),
+                            new(nameof(PlayerScore.LastName).Underscore(), playerScore.LastName),
+                            new(
+                                nameof(PlayerScore.UpdatedAt).Underscore(),
+                                playerScore.UpdatedAt.ToString(CultureInfo.InvariantCulture)
+                            ),
+                            new(
+                                nameof(PlayerScore.CreatedAt).Underscore(),
+                                playerScore.CreatedAt.ToString(CultureInfo.InvariantCulture)
+                            )
+                        }
+                    );
+                }
             }
         }
 
+        // in this case our primary database is postgres and should fill before cache
         if (_leaderBoardOptions.UseReadCacheAside)
         {
             if (!_leaderBoardDbContext.PlayerScores.Any())
